Added .dockerignore
docker-compose.yml:
restart: I am for fail fast
Moved the Dockerfile to the root
Disabled the volumes for prod, but re-enable if dev
ports value should be in double quotes
The test has no place in a separate container. It should be done in CI: Build the image, Test the code inside image in a separate temporary container, upload the image to a registry (hub.docker, or AWS ECR). Then CD: Deploy the image from the registry into a container in the prod env.

In Dockerfile the steps should be: copy package json, npm i, workdir, then copy the source files.